更改

+++ src/hybridcompiler/compiler.js
@@ -59,7 +59,7 @@ class Compiler{
 
         // 3. 对assets文件夹进行复制
         const assetsPath = path.join(cwd, 'assets');
-        if(fs.lstatSync(assetsPath).isDirectory()){
+        if(fs.existsSync(assetsPath) && fs.lstatSync(assetsPath).isDirectory()){
             fs.copySync(assetsPath, path.join(productionPath, 'assets'));
         }
